requirements.txt
# SSVEP BCI Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# EEG acquisition
brainflow>=5.0.0

# LSL and XDF for calibration data streaming and storage
pylsl>=1.16.0
pyxdf>=1.16.0

# Optional: Serial port detection for Cyton
pyserial>=3.5

# Optional: Advanced EEG analysis and debugging
mne>=1.0.0

# GUI (Phase 2)
PyQt6>=6.0.0
pyqtgraph>=0.13.0

# Development/testing
pytest>=7.0.0
